creates actions/searches and updates the imports
