+++ src/state/actions/index.js
@@ -13,12 +13,7 @@ export const REMOVE_BOOKMARKS = 'REMOVE_BOOKMARKS';
 export const SAVE_BOOKMARKS = 'SAVE_BOOKMARKS';
 export const THUMBNAIL = 'THUMBNAIL';
 
-export const CURRENT_SEARCH = 'CURRENT_SEARCH';
 export const LIST_VIEW = 'LIST_VIEW';
-export const SET_SEARCH_QUERY = 'SET_SEARCH_QUERY';
-export const SEARCH = 'SEARCH';
-
-export const SEARCH_BAR = 'SEARCH_BAR';
 
 export const SET_DOCS = 'SET_DOCS';
 
@@ -40,43 +35,11 @@ export const getDocs = authState => async dispatch => {
       dispatch({ type: SET_DOCS, payload: Response });
     } else {
       dispatch({ type: FINISH_FETCH });
-      dispatch({ type: SEARCH_BAR });
+      // dispatch({ type: SEARCH_BAR });
     }
   } catch (err) {
     console.log('GETDOCS ERROR', err);
   }
-};
-
-export const searchDocs = (search, authState, page, pageSize) => dispatch => {
-  dispatch({ type: START_FETCH });
-  dispatch({ type: SET_SEARCH_QUERY, payload: search });
-  getDSData(
-    `/search?query=${search}&page_number=${page -
-      1}&results_per_page=${pageSize}`,
-    authState
-  )
-    .then(data => {
-      console.log(data);
-      if (data.Response.length === 0) {
-        alert('No search results');
-        dispatch({ type: FINISH_FETCH });
-      } else {
-        dispatch({ type: SET_DOCS, payload: data });
-        dispatch({ type: SEARCH, payload: search });
-      }
-    })
-    .catch(console.error);
-};
-
-export const setCurrentSearch = (
-  currentSearch,
-  currentPage,
-  pageSize
-) => dispatch => {
-  dispatch({
-    type: CURRENT_SEARCH,
-    payload: { currentSearch, currentPage, pageSize },
-  });
 };
 
 export const saveBookmarks = (authState, bookmarkId) => async dispatch => {
